+++ src/MainPage/MainNav/MainNav.jsx
@@ -1,5 +1,5 @@
 import classes from './MainNav.module.css';
-import NavLink from './NavLink/NavLink';
+import MainNav__Link from '../MainNav__Link/MainNav__Link';
 import sprite from 'img/icon/sprite.svg';
 
 const MainNav = () => {
@@ -9,9 +9,9 @@ const MainNav = () => {
         <use xlinkHref={`${sprite}#icon-burger`} />
       </svg>
       <ul className={classes.navList}>
-        <NavLink linkName={'Главное'} />
-        <NavLink linkName={'Мой плейлист'} />
-        <NavLink linkName={'Выйти'} />
+        <MainNav__Link linkName={'Главное'} />
+        <MainNav__Link linkName={'Мой плейлист'} />
+        <MainNav__Link linkName={'Выйти'} />
       </ul>
     </nav>
   );
